ai-agent/requirements.txt
kubernetes==30.1.0
openai
python-dotenv